Moving around items. Creating Area, Room, Sector objects. Migration for Area.
