+++ db/migrate/20171226185648_create_area.rb
@@ -1,0 +1,7 @@
+class CreateArea < ActiveRecord::Migration[5.1]
+  def change
+    create_table :areas do |t|
+      t.string :name
+    end
+  end
+end
