+++ db/schema.rb
@@ -10,10 +10,14 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171225041035) do
+ActiveRecord::Schema.define(version: 20171226185648) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
+
+  create_table "areas", force: :cascade do |t|
+    t.string "name"
+  end
 
   create_table "players", force: :cascade do |t|
     t.string "name"
