+++ lib/models/area.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+require "active_record"
+
+class Area < ActiveRecord::Base
+end
