+++ lib/models/player.rb
