+++ lib/models/room.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+require "active_record"
+
+class Room < ActiveRecord::Base
+end
