+++ lib/sequence_loops/creation_process.rb
