+++ lib/sequence_loops/login_process.rb
