+++ spec/lib/models/area_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+require "spec_helper"
+
+RSpec.describe Area, type: [:models] do
+  let(:subject) { described_class }
+
+end
