+++ spec/lib/models/player_spec.rb
