+++ spec/lib/sequence_loops/creation_process_spec.rb
