+++ spec/lib/sequence_loops/login_process_spec.rb
